Added Archive Feature
Added feature that appends an archive of all references ever generated, archive can be searched by typing any element of reference e.g. author name or title and then list of best to worst matches will be displayed, choose the match of the source you are using and it will append it to the reference list.